refactor(vm): split program, value, and dispatch internals
